updated codebase to use 2017 toolchain

+++ RobotCasserole2016/src/org/usfirst/frc/team1736/robot/Pneumatics.java
@@ -28,7 +28,7 @@ public class Pneumatics {
 		return compressor.getPressureSwitchValue();
 	}
 	
-	public static float getCurrent(){
+	public static double getCurrent(){
 		return compressor.getCompressorCurrent();
 	}
